Remove need for dynamic output path in webpack config.
The output path changes depending on whether we are building a test file
or a result file. Testfiles must be placed in the directory that karma is
watching. Result files must not be placed in that directory. This change
makes the entrypoint name be the path differentiator instead of the
output.path. This is needed because I want to use the same webpack config
for both the test build and the entrypoint build.